Backend/requirements.txt
langchain 
langchain-community 
python-dotenv 
numpy 
pandas 
plotly 
fastapi 
pydantic
psycopg2-binary